The Go version now produces the same 4199 values as the C++ version for the word "emergent".
 Days of debugging to find that I had translated
   "if (currentDeltas[j]) {" into "if curDeltas[j] > 0.0"
    rather than "if curDeltas[j] != 0 {"
The problem was buried inside a couple of loops and other conditionals
and only changed a subset of the 4199 values scattered in the results.

To facilitate debugging I changed the way I was handling random number generation in both the C++ and Go code
so that I was working with the same random sequence in both versions.